Split resolver validation post pass
